+++ plugins/autocast/src/autocast.plugin.ts
@@ -1,51 +1,65 @@
 import api from '@flatfile/api'
-import { FlatfileListener, FlatfileEvent } from '@flatfile/listener'
-import { BulkRecordHook } from '@flatfile/plugin-record-hook'
 import { FlatfileRecord, TPrimitive } from '@flatfile/hooks'
+import { FlatfileEvent, FlatfileListener } from '@flatfile/listener'
+import { BulkRecordHook } from '@flatfile/plugin-record-hook'
+import { logInfo } from '@flatfile/util-common'
 
 export function autocast(
-  sheetId: string,
+  sheetSlug: string,
+  fieldFilters?: string[],
   options?: {
     chunkSize?: number
     parallel?: number
+    debug?: boolean
   }
 ) {
-  return async (client: FlatfileListener) => {
-    client.on('commit:created', { sheetId }, async (event: FlatfileEvent) => {
-      return BulkRecordHook(event, callback, options)
-    })
-  }
-}
+  return async (listener: FlatfileListener) => {
+    listener.on(
+      'commit:created',
+      { sheetSlug },
+      async (event: FlatfileEvent) => {
+        return await BulkRecordHook(
+          event,
+          async (records: FlatfileRecord[]) => {
+            const sheetId = event.context.sheetId
+            const sheet = await api.sheets.get(sheetId)
+            if (!sheet) {
+              logInfo('@flatfile/plugin-autocast', 'Failed to fetch sheet')
+              return
+            }
 
-async function callback(records: FlatfileRecord[], event: FlatfileEvent) {
-  const { sheetId } = event.context
-  const sheet = await api.sheets.get(sheetId)
-  if (!sheet) {
-    console.log(`Failed to fetch sheet with id: ${sheetId}`)
-    return
-  }
+            const castableFields = sheet.data.config.fields.filter((field) =>
+              fieldFilters
+                ? fieldFilters.includes(field.key)
+                : field.type !== 'string'
+            )
+            records.forEach((record) => {
+              castableFields.forEach((field) => {
+                const originalValue = record.get(field.key)
+                const caster = CASTING_FUNCTIONS[field.type]
 
-  const castableFields = sheet.data.config.fields.filter(
-    (field) => field.type !== 'string'
-  )
-
-  records.forEach((record) => {
-    castableFields.forEach((field) => {
-      const originalValue = record.get(field.key)
-      const caster = CASTING_FUNCTIONS[field.type]
-
-      if (originalValue && caster && typeof originalValue !== field.type) {
-        record.computeIfPresent(field.key, caster)
-
-        if (originalValue === record.get(field.key)) {
-          record.addError(
-            field.key,
-            `Failed to cast '${originalValue}' to '${field.type}'`
-          )
-        }
+                if (
+                  originalValue &&
+                  caster &&
+                  typeof originalValue !== field.type
+                ) {
+                  try {
+                    record.computeIfPresent(field.key, caster)
+                  } catch (e) {
+                    record.addError(
+                      field.key,
+                      e.message || 'Failed to cast value'
+                    )
+                  }
+                }
+              })
+            })
+          },
+          options
+        )
       }
-    })
-  })
+    )
+  }
 }
 
 const CASTING_FUNCTIONS: {
@@ -57,7 +71,9 @@ const CASTING_FUNCTIONS: {
 }
 
 export function castNumber(value: TPrimitive): TPrimitive {
-  if (typeof value === 'string') {
+  if (typeof value === 'number') {
+    return value
+  } else if (typeof value === 'string') {
     const strippedValue = value.replace(/,/g, '')
     if (!isNaN(Number(strippedValue))) {
       const num = Number(strippedValue)
@@ -66,13 +82,15 @@ export function castNumber(value: TPrimitive): TPrimitive {
       }
     }
   }
-  return value
+  throw new Error(`Failed to cast '${value}' to 'number'`)
 }
 
 export const TRUTHY_VALUES = ['1', 'yes', 'true', 'on', 't', 'y', 1]
 export const FALSY_VALUES = ['-1', '0', 'no', 'false', 'off', 'f', 'n', 0, -1]
 export function castBoolean(value: TPrimitive): TPrimitive {
-  if (typeof value === 'string' || typeof value === 'number') {
+  if (typeof value === 'boolean') {
+    return value
+  } else if (typeof value === 'string' || typeof value === 'number') {
     if (value === '') {
       return null
     }
@@ -84,7 +102,7 @@ export function castBoolean(value: TPrimitive): TPrimitive {
       return false
     }
   }
-  return value
+  throw new Error(`Failed to cast '${value}' to 'boolean'`)
 }
 
 export function castDate(value: TPrimitive): TPrimitive {
@@ -94,5 +112,5 @@ export function castDate(value: TPrimitive): TPrimitive {
       return date.toUTCString()
     }
   }
-  return value
+  throw new Error(`Failed to cast '${value}' to 'date'`)
 }