+++ plugins/autocast/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @flatfile/plugin-autocast
 
+## 0.2.2
+
+### Patch Changes
+
+- 3e28c51: Fix for date casting
+
+## 0.2.1
+
+### Patch Changes
+
+- ccbbbc9: Fix sheet filtering options to require sheetSlug.
+
+## 0.2.0
+
+### Minor Changes
+
+- 966cf91: This release introduces field filters, and enables filtering by either sheetId or sheetSlug.
+
 ## 0.1.0
 
 ### Minor Changes